+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## v0.10.4 (2025-05-05)
  - swap coords for CRS + special projections instead of SRS
+ - return at least the default german basemap
 
 ## v0.10.3 (2025-05-02)
 - technical release without any changes